requirements.txt
# Dependências principais da GimnTools
# Processamento numérico
numpy>=1.20.0
scipy>=1.7.0

# Compilação just-in-time para performance
numba>=0.54.0

pandas>=1.3.0
awkward>=1.2.0
uproot>=5.0.0

# Visualização e plotagem
matplotlib>=3.3.0

# Processamento de imagens médicas
SimpleITK>=2.1.0
itk>=5.2.0
scikit-image>=0.18.0

# Manipulação de imagens
Pillow>=8.0.0

# Armazenamento de dados
h5py>=3.1.0

# Interface de usuário
tqdm>=4.60.0

# Dependências opcionais para desenvolvimento
# Para instalar: pip install -r requirements.txt -r requirements-dev.txt

# Dependências para análise de dados avançada (opcional)
pandas>=1.3.0
seaborn>=0.11.0

# Notebooks Jupyter
jupyter>=1.0.0
ipywidgets>=7.6.0
jupyterlab>=3.0.0

# Visualização avançada
plotly>=5.0.0
seaborn>=0.11.0
